requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.0
gTTS==2.3.2
Pillow==10.0.0
requests==2.31.0
werkzeug==2.3.7
beautifulsoup4==4.12.2
